trying add property to node.

+++ lib/cocos2d-x/core/platform/CCEGLView.h
@@ -43,7 +43,7 @@ public:
 	virtual void swapBuffers();
 	virtual void setIMEKeyboardState(bool bOpen);
 	virtual void setFrameSize(const CCSize& size);
-	virtual void setFrameSize(int w, int h) { setFrameSize(CCSize(w, h)); }
+	virtual void setFrameSize(int w, int h) { setFrameSize(CCSize((float)w, (float)h)); }
 	virtual const CCSize& getFrameSize();
 	virtual const CCSize& getDrawableSize();
